Simplify initial TSC structure
During yesterday's meeting, the board discussed the governance mechanics specified in the TSC charter and bylaws. The current structure places heavy emphasis on wide representation and creating many opportunities for people to serve in governance roles. These are good goals, but they also have overhead costs — both higher process complexity and the increased coordination overhead of larger groups.

Moreover, while the board and TSC serve necessary supervisory functions, we don't want to set them up as the primary vehicles for participating in or keeping informed about the substantive activities of the BA. If we attempt to size these bodies to accommodate all of the BA's most prominent leaders, we risk a dynamic where they become the "room where it happens", limiting participation from other corners and creating further pressure to enlarge the groups.

The consensus was therefore to adjust our governance to be simpler and lighter-weight, in terms of both process and composition. It's possible we overshoot, but we'll retain the ability to add back some of these affordances should the need arise.

Here's a brief summary of the major proposed changes:
* Merging the elections for member directors, at-large directors, and TSC delegates into a single process overseen by the board and managed by the ED.
* Recomposing the TSC as 3 elected delegates along with one appointed delegate from each core project.
* Eliminating the separate role of the RC chair and delegating its responsibilities to the TSC chair.
* Specifying that the TSC chair need not undergo a formal election if they are unopposed, and that any newly-added delegates get a voice in chair selection.
* Reducing the number of member directors on the board from 8 to 4 (for a total of 7 directors), and setting up an annual election cadence.

These require changes to both the TSC charter and to the bylaws. This PR contains the proposed charter changes.