proposal: jco-transpile hosted project
This commit proposes the @bytecodealliance/jco-transpile project, a
subproject of @bytecodealliance/jco that is meant to contain only the
usable bits of jco for WebAssembly component transpilation.

The creation of this project enables downstream projects to pull in
fewer dependencies when utilizing transpilation functionality of jco
while also freeing up jco to become a batteries-included multi-tool.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>