Update the RC Program to not require voting
The TSC has received feedback from multiple people that they're not comfortable accepting nominations to become a Recognized Contributor, because they don't want to vote in elections to the TSC or Board of Directors.

At the last TSC meeting, we agreed to drop this requirement. This commit changes the TSC charter accordingly, and tweaks some of the language about the RC Program in doing so.